A example of the leech attack
Leech is being used by a new enemy, called "the overgrowth". Designed to be just an example (brick take a look), it shines in its (very rushed and not finalized) very unique abilities. Simply put, if it ever gets low on hp, and it has enough mana, it attempts to leech the player. 
If it runs low on mana, or has enough hp, it will attempt to "steal" Your mana, making it absurdly balanced as it is only as strong as the opposing player.