This might be stupid
Attempted to make a log that says if the player uses a spell so that the enemy entity can do damage to it and punish the player if he tries to do anything special. I have no clue about the code, but based on what I figured, it works like as so.